ton-o messengers/defaults added

+++ src/main/java/moe/maple/api/script/model/object/field/LifeObject.java
@@ -29,7 +29,8 @@ public interface LifeObject extends FieldedObject {
 
     // =================================================================================================================
 
-    int getHealthCurrent();
+    long getHealthCurrent();
+
     int getManaCurrent();
 
     int getHealthMax();
